# docs/admin-executor-openapi.yaml
openapi: 3.0.3
info:
  title: Admin Executor API
  version: 0.1.0
  description: |
    Mock admin API for plan/org/tier/credit management.
    Use /api/auth to issue a bearer token, then call admin endpoints.
servers:
  - url: http://localhost:3000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    ErrorResponse:
      type: object
      required: [ok, message]
      properties:
        ok:
          type: boolean
          example: false
        message:
          type: string
          example: unauthorized
...
    Actor:
      type: object
      required: [id, scopes]
      properties:
        id:
          type: string
          example: U123
        scopes:
          type: array
          items:
            type: string
          example: ["PERSONAL_PLAN_ADMIN"]
    AdminCommandRequest:
      type: object
      required: [rawText, actor]
      properties:
        rawText:
          type: string
          example: pro-plan grant --userId=user123
        actor:
          $ref: "#/components/schemas/Actor"
        credentials:
          type: object
          properties:
            username:
              type: string
            password:
              type: string
    AdminCommandResponse:
      type: object
      required: [ok, message]
      properties:
        ok:
          type: boolean
        message:
          type: string
    UserSummary:
      type: object
      required: [ok, userId, plan, orgId]
      properties:
        ok:
          type: boolean
        userId:
          type: string
        plan:
          type: string
        orgId:
          type: string
          nullable: true
...
    OrgSummary:
      type: object
      required: [ok, orgId, tier, credit]
      properties:
        ok:
          type: boolean
        orgId:
          type: string
        tier:
          type: string
        credit:
          type: integer
    PlanGrantRequest:
      type: object
      properties:
        plan:
          type: string
          example: pro
...
    TierUpdateRequest:
      type: object
      required: [tier]
      properties:
        tier:
          type: string
          example: pro
...
paths:
...
  /api/admin:
    post:
      tags: [admin]
      summary: Execute admin command gateway
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCommandRequest"
      responses:
        "200":
          description: Command executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCommandResponse"
        "400":
          description: Invalid command
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/admin/users/{userId}/plan/grant:
    post:
      tags: [users]
      summary: Grant or update user plan
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanGrantRequest"
      responses:
        "200":
          description: Plan updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSummary"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/admin/orgs/{orgId}/tier/update:
    post:
      tags: [orgs]
      summary: Update org tier
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: orgId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TierUpdateRequest"
      responses:
        "200":
          description: Tier updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgSummary"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...